requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
python-pptx>=0.6.21
python-docx>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
